backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.0.76
Pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
python-dotenv==1.0.0
pytest==7.4.2
black==23.7.0
flake8==6.0.0